# backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "윤정 법률사무소 API",
    "version": "1.0.0",
    "description": "윤정 법률사무소 홈페이지 백엔드 REST API. 모든 응답은 { data, error, meta } 형식입니다."
  },
  "servers": [
    { "url": "/api", "description": "기본 API 서버" }
  ],
...
  "paths": {
...
    "/admin-users/login": {
      "post": {
        "tags": ["AdminUsers"],
        "summary": "관리자 로그인",
        "description": "아이디/비밀번호로 인증 후 Bearer 토큰을 반환합니다. IP당 15분간 10회 제한.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": { "type": "string", "description": "관리자 아이디" },
                  "password": { "type": "string", "description": "비밀번호" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "로그인 성공",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "token": { "type": "string", "description": "Bearer 인증 토큰" },
                            "user": {
                              "type": "object",
                              "properties": {
                                "id": { "type": "string" },
                                "username": { "type": "string" },
                                "name": { "type": "string" },
                                "role": { "type": "string" },
                                "email": { "type": "string", "nullable": true }
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": {
            "description": "비활성화된 계정",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponse" }
              }
            }
          }
        }
      }
    },
...
    "/admin-users/me": {
      "get": {
        "tags": ["AdminUsers"],
        "summary": "현재 로그인된 관리자 정보",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "사용자 정보 (passwordHash 제외)",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/AdminUser" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/admin-users": {
      "get": {
        "tags": ["AdminUsers"],
        "summary": "관리자 사용자 목록 (관리자)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "관리자 사용자 목록",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/AdminUser" }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["AdminUsers"],
        "summary": "관리자 사용자 생성 (관리자)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password", "name"],
                "properties": {
                  "username": { "type": "string", "description": "아이디" },
                  "password": { "type": "string", "description": "비밀번호" },
                  "name": { "type": "string", "description": "이름" },
                  "role": { "type": "string", "enum": ["admin", "editor"], "default": "editor", "description": "역할" },
                  "email": { "type": "string", "nullable": true, "description": "이메일" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "생성된 사용자",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/AdminUser" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "409": {
            "description": "아이디 중복",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponse" }
              }
            }
          }
        }
      }
    },
    "/admin-users/{id}": {
      "patch": {
        "tags": ["AdminUsers"],
        "summary": "관리자 사용자 수정 (관리자)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" }, "description": "사용자 ID (UUID)" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "password": { "type": "string", "description": "새 비밀번호 (변경 시)" },
                  "name": { "type": "string" },
                  "role": { "type": "string", "enum": ["admin", "editor"] },
                  "email": { "type": "string", "nullable": true },
                  "isActive": { "type": "boolean", "description": "활성화 여부" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "수정된 사용자",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/AdminUser" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["AdminUsers"],
        "summary": "관리자 사용자 비활성화 (소프트 삭제, 관리자)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" }, "description": "사용자 ID (UUID)" }
        ],
        "responses": {
          "200": {
            "description": "비활성화된 사용자",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/AdminUser" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "관리자 인증 토큰 (POST /admin-users/login에서 발급)"
      },
      "portalAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-portal-token",
        "description": "포털 세션 토큰 (POST /portal/login에서 발급, TTL 24시간)"
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "description": "모든 API의 공통 응답 형식",
        "properties": {
          "data": { "description": "응답 데이터 (성공 시)", "nullable": true },
          "error": { "type": "string", "description": "에러 메시지 (실패 시)", "nullable": true },
          "meta": { "type": "object", "description": "부가 정보 (페이지네이션 등)", "nullable": true }
        }
      },
...
      "AdminUser": {
        "type": "object",
        "description": "관리자 사용자 (passwordHash 제외)",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "username": { "type": "string" },
          "name": { "type": "string" },
          "role": { "type": "string", "enum": ["admin", "editor"] },
          "email": { "type": "string", "nullable": true },
          "isActive": { "type": "integer", "description": "1=활성, 0=비활성" },
          "lastLoginAt": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "잘못된 요청",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                { "$ref": "#/components/schemas/ApiResponse" },
                {
                  "type": "object",
                  "properties": {
                    "data": { "type": "null" },
                    "error": { "type": "string", "example": "필수 항목이 누락되었습니다" }
                  }
                }
              ]
            }
          }
        }
      },
      "Unauthorized": {
        "description": "인증 실패",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                { "$ref": "#/components/schemas/ApiResponse" },
                {
                  "type": "object",
                  "properties": {
                    "data": { "type": "null" },
                    "error": { "type": "string", "example": "인증이 필요합니다" }
                  }
                }
              ]
            }
          }
        }
      },
      "NotFound": {
        "description": "리소스를 찾을 수 없음",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                { "$ref": "#/components/schemas/ApiResponse" },
                {
                  "type": "object",
                  "properties": {
                    "data": { "type": "null" },
                    "error": { "type": "string", "example": "리소스를 찾을 수 없습니다" }
                  }
                }
              ]
            }
          }
        }
      }
    }
  }
}